# openapi.yaml
openapi: 3.0.0
info:
  title: API Agenda Personal
  version: 1.0.0
servers:
  - url: https://gpt-agenda.onrender.com
paths:
  /crear_evento:
    post:
      summary: Crear evento
      operationId: crearEvento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summary:
                  type: string
                start:
                  type: string
                  description: Fecha y hora de inicio (YYYY-MM-DD HH:MM)
                end:
                  type: string
                  description: Fecha y hora de fin (YYYY-MM-DD HH:MM)
      responses:
        '200':
          description: Evento creado

  /leer_agenda:
    get:
      summary: Leer agenda
      operationId: leerAgenda
      parameters:
        - name: rango
          in: query
          required: true
          schema:
            type: string
            enum: [diario, semanal, mensual]
      responses:
        '200':
          description: Lista de eventos

  /eliminar_evento/{id_evento}:
    delete:
      summary: Eliminar evento
      operationId: eliminarEvento
      parameters:
        - name: id_evento
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Evento eliminado

  /modificar_evento/{id_evento}:
    patch:
      summary: Modificar evento
      operationId: modificarEvento
      parameters:
        - name: id_evento
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                summary:
                  type: string
                start:
                  type: string
                end:
                  type: string
      responses:
        '200':
          description: Evento modificado

  /sugerir_horario:
    get:
      summary: Sugerir horario disponible
      operationId: sugerirHorario
      parameters:
        - name: duracion
          in: query
          required: true
          schema:
            type: integer
            description: Duración en minutos
      responses:
        '200':
          description: Horario sugerido